+++ src/bindings/buffer/mod.rs
@@ -1,5 +1,6 @@
 pub mod blas;
 pub mod index;
+pub mod mutability;
 pub mod storage;
 pub mod uniform;
 pub mod vertex;