+++ src/misc/mod.rs
@@ -1,4 +1,4 @@
+pub mod camera;
 pub(crate) mod default_buffer;
 pub(crate) mod ids;
-pub mod mutability;
 pub(crate) mod thread_ptr;